+++ backend/src/chat/channel.module.ts
@@ -5,17 +5,17 @@ import { ChannelController } from './channel.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from 'src/entity/User';
-
+import { UserService } from 'src/user.service';
 
 @Module({
-  imports: 
+  imports:
   [TypeOrmModule.forFeature([ User ]),
   JwtModule.register({
     global: true,
     secret: process.env.JWT_SECRET,
     signOptions: { expiresIn: '1d'}
   })],
-  providers: [ChannelGateway, ChannelService],
+  providers: [ChannelGateway, ChannelService, UserService],
   controllers: [ChannelController]
 })
 export class ChannelModule {}
