+++ backend/src/user.service.ts
@@ -1,11 +1,46 @@
 import { Injectable } from '@nestjs/common';
 import { UserInfo, IntraId} from 'src/type/userType';
-import { SocketId } from 'src/type/socketType';
+import { SocketId, Socket } from 'src/type/socketType';
+import { JwtService } from '@nestjs/jwt';
+import { PnPayloadDto } from 'src/chat/channel.dto';
+import * as cookie from 'cookie';
 
 @Injectable()
 export class UserService {
+    constructor(private readonly jwtService: JwtService) {}
+
     userMap = new Map<number, UserInfo>();
     idMap = new Map<SocketId, IntraId>();
+
+    checkPnJwt(client: Socket) {
+      console.log('in the checktPnJwt');
+      const cookies = client.handshake.headers.cookie;
+      console.log('cookies', cookies);
+      if (!cookies) {
+        console.error('Cookies not found');
+        return false;
+      }
+      const pnJwtCookie = cookie.parse(cookies)['pn-jwt'];
+
+      if (!pnJwtCookie) {
+        console.error('JWT not found');
+        return false;
+      }
+      try {
+        const payload: PnPayloadDto = this.jwtService.verify<PnPayloadDto>(pnJwtCookie);
+        if (payload.exp * 1000 < Date.now()) {
+          console.error('JWT expired');
+          return false;
+        }
+        client.intraId = payload.intraId;
+        console.log('client', client.id);
+        console.log('client.intraId', client.intraId);
+      } catch (err) {
+        console.error('JWT verification failed', err);
+        return false;
+      }
+      return true;
+    }
 
     addUser(userInfo: UserInfo, clientId: SocketId) {
       this.userMap.set(userInfo.intraId, userInfo);
